Polished up the team page as well fixed the team link
