+++ client/src/App.js
@@ -9,7 +9,6 @@ import Dashboard from './components/Dashboard';
 import Project from './components/Project';
 import Team from './components/Team';
 import Issues from './components/Issues';
-import MeetTheDevs from './components/MeetTheDevs';
 
 function App() {
   return (
@@ -23,7 +22,7 @@ function App() {
             <Route path='/project' element={<Project />} />
             <Route path='/team' element={<Team />} />
             <Route path='/issues' element={<Issues />} />
-            <Route path='/meetTheDevs' element={<MeetTheDevs />} />
+            <Route path='/team' element={<Team />} />
           </Routes>
         </main>
         <Footer />
